A script to extract thermochemistry data.
The script also contains functions to extract vibrational frequencies from an Orca output file. If used as "__main__" use as follows:

$ python thermodata_orca_to_csv.py <filename.out> <path_to_directory> <temperature>

The script looks for all files with the name "filename.out" in "path_to_directory", extracts the data that was calculated at the specified "temperature" (in K), and creates a csv that contains the extracted data. One row for each file.